ndk/native_window: Add `format()` getter and `set_buffers_geometry()` setter
Android's `NativeWindow` allow querying the format through
`ANativeWindow_getFormat` and setting it through
`ANativeWindow_setBuffersGeometry` together with the dimensions to use
for the buffer.
