+++ ndk/CHANGELOG.md
@@ -8,6 +8,7 @@
 - **Breaking:** audio: Rename from `aaudio` to `audio` and drop `A` prefix. (#273)
 - Implement `HasRawWindowHandle` directly on `NativeWindow`. (#274)
 - **Breaking:** native_activity: Replace `CStr` return types with `Path`. (#279)
+- native_window: Add `format()` getter and `set_buffers_geometry()` setter. (#276)
 
 # 0.6.0 (2022-01-05)
